11/3/15 - total area is now total area after densityConnector(), statsign uses 4.5^mul

+++ WebDerm/src/ShadeShape/Labels/labels.h
@@ -20,7 +20,7 @@ private:
 	map<String,int> labelShapeNumMap;
 	map<String,int> labelPrevShapeNumMap;
 	map<String,int> labelShadeLevelMap;
-	map<String, vector<int> > labelStatSignMap;
+	map<String, vector<float> > labelStatSignMap;
 	String labelName;
 	int labelTotalArea;
 	void calcTotalArea();
@@ -50,8 +50,8 @@ public:
 	bool isShapeShifted(String label);
 	int getShapeNum(String label);
 	int getPrevShapeNum(String label);
-	vector<int> getStatSign(String label);
-	map<String,vector<int>>& getStatSignMap();
+	vector<float> getStatSign(String label);
+	map<String,vector<float>>& getStatSignMap();
 	static void printCompareLabels(Labels &labels1, Labels &labels2, int markShifted=0);
 	static void writeCompareLabels(Labels &labels1, Labels &labels2, int markShifted=0);
 	static void printCompareStatSign(Labels &labels1, Labels &labels2, String label);
